Add a separate theme preference for terminal output
This is useful because a lot of people seem to prefer a dark theme for
terminal output pastes, but a light theme for normal pastes.

This makes monokai the default for terminal output pastes, and keeps
"default" as the default for regular pastes. The theme preference is now
separated out: you can have a theme configured for regular pastes, and a
separate theme configured for terminal output pastes.